+++ uilayer-core/src/test/java/net/avh4/framework/uilayer/test/TestGraphicsOperations.java
@@ -19,7 +19,7 @@ public class TestGraphicsOperations implements GraphicsOperations {
     }
 
     @Override
-    public void drawText(String text, float leftX, float baselineY, Font font, int argbColor) {
+    public void drawText(String text, double leftX, double baselineY, Font font, int argbColor) {
         sb.append("Text: \"").append(text).append("\" ");
         sb.append(leftX).append(", ");
         sb.append(baselineY).append(" ");
